break out components
